# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MASHALL NGO website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface VolunteerInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  message: string;
}

export interface FormResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}