# lib/api-zod/src/generated/types/formResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MASHALL NGO website API
 * OpenAPI spec version: 0.1.0
 */

export interface FormResponse {
  success: boolean;
  message: string;
}