# lib/api-zod/src/generated/types/volunteerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MASHALL NGO website API
 * OpenAPI spec version: 0.1.0
 */

export interface VolunteerInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  message: string;
}